requirements-dev.txt
Telethon==1.28.5
numpy==1.24.3
pandas==2.0.3
pymongo==4.3.3
colorama==0.4.6
Markdown==3.4.3
schedule==1.2.0
PyYAML==5.3.1
requests==2.31.0